+++ examples/plot_aia_overplotting.py
@@ -19,6 +19,7 @@ import sunpy.io.fits
 
 import pfsspy
 import pfsspy.coords as coords
+import pfsspy.tracing as tracing
 
 
 ###############################################################################
@@ -117,10 +118,11 @@ output = pfsspy.pfss(input)
 # Trace field lines from the footpoints defined above. :mod:`pfsspy.coords`
 # is used to convert the s, phi cooridnates into the cartesian coordinates that
 # are needed by the tracer.
+tracer = tracing.PythonTracer(atol=1e-8)
 flines = []
 for s, phi in zip(s.ravel(), phi.ravel()):
     x0 = np.array(pfsspy.coords.strum2cart(0.01, s, phi))
-    flines.append(output.trace(x0, atol=1e-8))
+    flines += tracer.trace(x0, output)
 
 ###############################################################################
 # Plot the input GONG magnetic field map, along with the traced mangetic field
@@ -128,8 +130,9 @@ for s, phi in zip(s.ravel(), phi.ravel()):
 fig, ax = plt.subplots()
 mesh = input.plot_input(ax)
 for fline in flines:
-    fline.representation_type = 'spherical'
-    ax.plot(fline.lon / u.deg, np.sin(fline.lat), color='black', linewidth=1)
+    fline.coords.representation_type = 'spherical'
+    ax.plot(fline.coords.lon / u.deg, np.sin(fline.coords.lat),
+            color='black', linewidth=1)
 
 ax.set_xlim(55, 65)
 ax.set_ylim(0.1, 0.25)
@@ -143,11 +146,8 @@ ax = plt.subplot(1, 1, 1, projection=aia)
 transform = ax.get_transform('world')
 aia.plot(ax)
 for fline in flines:
-    fline = fline.transform_to(aia.coordinate_frame)
-    Tx = fline.Tx.to(u.deg)
-    Ty = fline.Ty.to(u.deg)
-    ax.plot(Tx, Ty, transform=transform,
-            alpha=0.8, linewidth=1, color='black')
+    coords = fline.coords.transform_to(aia.coordinate_frame)
+    ax.plot_coord(coords, alpha=0.8, linewidth=1, color='black')
 
 ax.set_xlim(500, 900)
 ax.set_ylim(400, 800)